move validators to parsers
