+++ src/sentry/api/validators/__init__.py
@@ -1,10 +1,8 @@
 from .auth import *  # noqa: F401,F403
 from .broadcast import *  # noqa: F401,F403
 from .doc_integration import *  # noqa: F401,F403
-from .email import *  # noqa: F401,F403
 from .external_actor import *  # noqa: F401,F403
 from .integrations import *  # noqa: F401,F403
 from .notifications import *  # noqa: F401,F403
 from .project import *  # noqa: F401,F403
 from .project_codeowners import *  # noqa: F401,F403
-from .userrole import *  # noqa: F401,F403
